requirement.txt
torch
torchvision
numpy
matplotlib
scikit-learn
opencv-python
pyyaml
tqdm
